add publicly reusable StringIO object
