+++ lib/rspec_fixtures.rb
@@ -1,6 +1,7 @@
 require 'rspec_fixtures/extensions/file'
 
-require 'rspec_fixtures/stream_capturer'
+require 'rspec_fixtures/module_functions'
+require 'rspec_fixtures/stream'
 require 'rspec_fixtures/approval_handler'
 require 'rspec_fixtures/matchers/base'
 require 'rspec_fixtures/matchers/match_fixture'
@@ -8,3 +9,4 @@ require 'rspec_fixtures/matchers/output_fixture'
 require 'rspec_fixtures/matchers/raise_fixture'
 
 require 'rspec_fixtures/rspec_config'
+
