+++ lib/rspec_fixtures/module_functions.rb
@@ -1,0 +1,11 @@
+module RSpecFixtures
+  class << self
+    def stdout
+      @stdout ||= StringIO.new
+    end
+
+    def stderr
+      @stderr ||= StringIO.new
+    end
+  end
+end
