+++ spec/rspec_fixtures/stream_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
-describe CaptureStdout do
+describe Stream::Stdout do
   let(:proc) { Proc.new { print "hello world" } }
 
   it "captures stdout" do
-    expect(CaptureStdout.capture proc).to eq "hello world"
+    expect(described_class.capture proc).to eq "hello world"
   end
 end
 
-describe CaptureStderr do
+describe Stream::Stderr do
   let(:proc) { Proc.new { $stderr.print "hello world" } }
 
   it "captures stdout" do
-    expect(CaptureStderr.capture proc).to eq "hello world"
+    expect(described_class.capture proc).to eq "hello world"
   end
 end
